LayoutTests/platform/qt-mac/fast/multicol/vertical-lr/break-properties-expected.txt

PASS: 'break-before' is at (8, 218)
PASS: 'after-break' is at (8, 428)
